requirements.txt
pyautogen>=0.10
autogen-ext[openai]
python-dotenv
openai
streamlit
PyMuPDF
